chore(module-federation): clean up module-federation executors for react
