+++ CODEOWNERS
@@ -96,7 +96,6 @@ rust-toolchain @nrwl/nx-native-reviewers
 /packages/webpack/** @nrwl/nx-js-reviewers
 /e2e/webpack/** @nrwl/nx-js-reviewers
 /packages/rspack/** @nrwl/nx-js-reviewers
-/packages/rspack/src/utils/module-federation @nrwl/nx-js-reviewers
 /e2e/rspack/** @nrwl/nx-js-reviewers
 /packages/esbuild/** @nrwl/nx-js-reviewers
 /e2e/esbuild/** @nrwl/nx-js-reviewers
