+++ packages/react/src/executors/module-federation-dev-server/lib/index.ts
@@ -1,0 +1,2 @@
+export * from './normalize-options';
+export * from './start-remotes';
