+++ packages/rspack/src/executors/module-federation-dev-server/module-federation-dev-server.impl.ts
@@ -1,11 +1,4 @@
-import {
-  ExecutorContext,
-  logger,
-  parseTargetString,
-  ProjectConfiguration,
-  readTargetOptions,
-  runExecutor,
-} from '@nx/devkit';
+import { ExecutorContext, logger } from '@nx/devkit';
 import {
   combineAsyncIterables,
   createAsyncIterable,
@@ -16,92 +9,8 @@ import { existsSync } from 'fs';
 import { extname, join } from 'path';
 import { startRemoteIterators } from '@nx/module-federation/src/executors/utils';
 import devServerExecutor from '../dev-server/dev-server.impl';
-import {
-  ModuleFederationDevServerOptions,
-  NormalizedModuleFederationDevServerOptions,
-} from './schema';
-
-function getBuildOptions(buildTarget: string, context: ExecutorContext) {
-  const target = parseTargetString(buildTarget, context);
-
-  const buildOptions = readTargetOptions(target, context);
-
-  return {
-    ...buildOptions,
-  };
-}
-
-async function startRemotes(
-  remotes: string[],
-  workspaceProjects: Record<string, ProjectConfiguration>,
-  options: Partial<
-    Pick<
-      ModuleFederationDevServerOptions,
-      'devRemotes' | 'host' | 'ssl' | 'sslCert' | 'sslKey' | 'verbose'
-    >
-  >,
-  context: ExecutorContext,
-  target: 'serve' | 'serve-static' = 'serve'
-) {
-  const remoteIters: AsyncIterable<{ success: boolean }>[] = [];
-
-  for (const app of remotes) {
-    const remoteProjectServeTarget = workspaceProjects[app].targets[target];
-    const isUsingModuleFederationDevServerExecutor =
-      remoteProjectServeTarget.executor.includes(
-        'module-federation-dev-server'
-      );
-
-    const configurationOverride = options.devRemotes?.find(
-      (
-        r
-      ): r is {
-        remoteName: string;
-        configuration: string;
-      } => typeof r !== 'string' && r.remoteName === app
-    )?.configuration;
-
-    const defaultOverrides = {
-      ...(options.host ? { host: options.host } : {}),
-      ...(options.ssl ? { ssl: options.ssl } : {}),
-      ...(options.sslCert ? { sslCert: options.sslCert } : {}),
-      ...(options.sslKey ? { sslKey: options.sslKey } : {}),
-    };
-    const overrides =
-      target === 'serve'
-        ? {
-            watch: true,
-            ...(isUsingModuleFederationDevServerExecutor
-              ? { isInitialHost: false }
-              : {}),
-            ...defaultOverrides,
-          }
-        : { ...defaultOverrides };
-
-    remoteIters.push(
-      await runExecutor(
-        {
-          project: app,
-          target,
-          configuration: configurationOverride ?? context.configurationName,
-        },
-        overrides,
-        context
-      )
-    );
-  }
-  return remoteIters;
-}
-
-function normalizeOptions(
-  options: ModuleFederationDevServerOptions
-): NormalizedModuleFederationDevServerOptions {
-  return {
-    ...options,
-    devRemotes: options.devRemotes ?? [],
-    verbose: options.verbose ?? false,
-  };
-}
+import { ModuleFederationDevServerOptions } from './schema';
+import { getBuildOptions, normalizeOptions, startRemotes } from './lib';
 
 export default async function* moduleFederationDevServer(
   schema: ModuleFederationDevServerOptions,
