+++ packages/rspack/src/executors/module-federation-ssr-dev-server/lib/index.ts
@@ -1,0 +1,2 @@
+export * from './normalize-options';
+export * from './start-remotes';
